Use sha256 instead of md5 for hashing items (#4)
* Use sha256 instead of md5 for hashing items

I am personally incredibly sorry to be sending this PR.

pyca/cryptography is looking at using pytest-shard to speed up our tests (looks like it'll work great!). One of our CI jobs tests against OpenSSL in FIPS mode.

FIPS 140 is a cryptographic standard from the US Government, it's very bad, almost comedically so. One of the things it does is ban md5. This is because md5 is broken for most security purposes.

It is no doubt extremely obvious that md5 is not being used in a security sensitive way here. Nevertheless, OpenSSL in FIPS mode errors when you try to use it. So I've sent this PR, which simply swaps it for SHA256, which the FIPS gods allow.

* Trigger CI

* Fix the tests

* Update test_pytest_shard.py

s/sha2hash/sha256hash

Co-authored-by: Adam Gleave <adam@gleave.me>